#2 Cast of bytes

+++ app/src/main/java/com/apulido/calculadora/MainActivity.java
@@ -37,8 +37,8 @@ public class MainActivity extends AppCompatActivity {
         text_view_result = (TextView)findViewById(R.id.Result);
 
         byte[] BEACommand = new byte[4];
-        BEACommand[0] = 0xFF;
-        BEACommand[1] = 0x05;
+        BEACommand[0] = (byte)0xFF;
+        BEACommand[1] = (byte)0x05;
         BEACommand[0] = 0xFF;
         BEACommand[1] = 0x05;
